+++ src/components/dashboard/SideBar/SharedProjectItem.tsx
@@ -1,0 +1,2 @@
+const SharedProjectItem = () => {};
+export default SharedProjectItem;
