+++ src/constants/dashboard/sideBar.ts
@@ -1,0 +1,4 @@
+export const PAGE_CONFIG = {
+  대시보드: "/dashboard",
+  프로젝트: "/dashboard/project",
+} as const;
